#1282 * Updated `cp.rx.go` documentation * Split up the files to make them easier to navigate and manage.
